Simplify passing files
Right now, handling the files passed to the child process is up
to the user of the package. This makes the package simpler, but harder
to use. Specifically, the user has to make sure they don't close
files / listeners prematurely. Implicitly, it's also the user's
responsibility to close files they are not using anymore.

Those two points make it cumbersome to use the package correctly.

Instead, add a bit more boilerplate which takes care of both of these issues.
This is the same approach used by FB's gracenet, but with different details.
The Fds struct tracks inherited but unused, and inherited but used files.
When a new file is added to Fds it is duplicated, to prevent the user from
prematurely closing the file.
File descriptors which were inherited but not used are closed when Ready() is
called, and ones that were used are closed when Stop() is called.