Unlink unused UNIX sockets
Right now, a UNIX socket is never unlinked from the file system,
even if an upgraded version of a service does not use it
anymore. This is because we need to prevent Go from unlinking
the UNIX socket by default.

Fix this by removing unused filesystem-bound sockets when Ready()
is called.