Allow opening connection with callbacks
Sometimes, it is useful to customize the listening connections, while
the `ListenConfig` object allows to do this to some extent, it has the
limitation to be shared by all connections managed by an Upgrader
object.

This is a problem if only some particular connections have to be customized
in a specific way.

This commit adds new methods where the user has to provide callbacks that
are responsible to open the connection when needed.