add function to return all inherited files
This is helpful in the use case of integrating with systemd socket activation.

When there is no parent, the app still need to scan or use actication.Files() to
find all the fds and call upgrader.AddFile() to track them. But when there is a
parent, we can access the fds from upgrader instance rather then by scanning
fds or calling actication.Files() both of which creates os.File objects for the fds.

Without this change, there would be two set of os.File objects that represent
the same fd set. One set of os.File is hold by the upgrader whie the another set
returning from fd scanning (or actication.Files()) is being used at other places.
This raise a risk of closing the same fd twice at different time which is really bad
when the fd get reused by other files during the two close calls.